refactor: move season lifecycle slice into core application
