+++ backend/src/core/application/ports/__init__.py
@@ -9,6 +9,7 @@ from .pena_query_port import PenaQueryPort
 from .pena_season_port import PenaSeasonPort
 from .player_profile_port import PlayerProfilePort
 from .registration_port import AdminRegistrationPort, UserRegistrationPort
+from .season_competition_port import SeasonCompetitionPort
 from .season_match_insights_port import SeasonMatchInsightsPort
 
 __all__ = [
@@ -23,6 +24,7 @@ __all__ = [
     "PenaQueryPort",
     "PenaSeasonPort",
     "PlayerProfilePort",
+    "SeasonCompetitionPort",
     "SeasonMatchInsightsPort",
     "UserRegistrationPort",
 ]
